% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSAnnealingGUI.r
\name{AnnPro}
\alias{AnnPro}
\title{Displays the trend of the ratio of core line intensities or Fit component intensities
as a function of the annealing temperature associated to a given XPSSample. Annealing temperature
is manually provided to the GUI.
A fit using an exponential decay function is then used to fit the data}
\usage{
AnnPro()
}
\description{
Displays the trend of the ratio of core line intensities or Fit component intensities
as a function of the annealing temperature associated to a given XPSSample. Annealing temperature
is manually provided to the GUI.
A fit using an exponential decay function is then used to fit the data
}
\examples{

\dontrun{
DepthPro()
}

}
