% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{AsymmGauss}
\alias{AsymmGauss}
\title{AsymmGauss function
Tail modified Gauss by an exponential function}
\usage{
AsymmGauss(x, h, mu, sigma, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{asym}{function asymmetry value}
}
\description{
AsymmGauss function
Tail modified Gauss by an exponential function
}
