% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{AsymmGaussLorentzProd}
\alias{AsymmGaussLorentzProd}
\title{AsymmGaussLorentzProd function
Asym Gaussian Lorentz cross product from UNIFIT Publication}
\usage{
AsymmGaussLorentzProd(x, h, mu, sigma, asym, lg)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{asym}{function asymmetry value}

\item{lg}{mix Gauss-Lorentz}
}
\description{
AsymmGaussLorentzProd function
Asym Gaussian Lorentz cross product from UNIFIT Publication
}
