% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{AsymmGaussVoigt}
\alias{AsymmGaussVoigt}
\title{AsymmGaussVoigt Asymmetric Gaussian-Voigt function see CasaXPS}
\usage{
AsymmGaussVoigt(x, h, mu, sigma, lg, asym, gv)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{lg}{mix Gauss-Lorentz}

\item{asym}{function asymmetry value}

\item{gv}{mix Gauss-Voigt}
}
\description{
AsymmGaussVoigt Asymmetric Gaussian-Voigt function see CasaXPS
}
