% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{AsymmVoigt}
\alias{AsymmVoigt}
\title{AsymmVoigt}
\usage{
AsymmVoigt(x, h, mu, sigma, lg, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{lg}{mix Gauss-Lorentz}

\item{asym}{function asymmetry value}
}
\description{
AsymmVoigt
}
