% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{DoniachSunjic}
\alias{DoniachSunjic}
\title{DoniachSunjic function
Correct version of the Doniach-Sunjic function (see Wertheim PRB 25(3), 1987, (1982))}
\usage{
DoniachSunjic(x, h, mu, sigmaDS, asym)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigmaDS}{function full width at half maximum}

\item{asym}{function asymmetry value}
}
\description{
DoniachSunjic function
Correct version of the Doniach-Sunjic function (see Wertheim PRB 25(3), 1987, (1982))
}
