% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{DoniachSunjicGaussTail}
\alias{DoniachSunjicGaussTail}
\title{DoniachSunjicGauss function
Doniach Sunjic function multiplied for a Gaussian broadening
corrected for an exponential decay on the low BE side}
\usage{
DoniachSunjicGaussTail(x, h, mu, sigmaDS, sigmaG, asym, tail)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigmaDS}{function full width at half maximum}

\item{sigmaG}{full width at half maximum of superimpossed Gaussian broadening}

\item{asym}{function asymmetry value}

\item{tail}{amplitude of the spectral tail on the low BE (high KE) side}
}
\description{
DoniachSunjicGauss function
Doniach Sunjic function multiplied for a Gaussian broadening
corrected for an exponential decay on the low BE side
}
