% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{DoniachSunjicTail}
\alias{DoniachSunjicTail}
\title{DoniachSunjic function
Version of the Doniach-Sunjic function (see Wertheim PRB 25(3), 1987, (1982))
Corrected for an exponential decay (tail) on the low BE side}
\usage{
DoniachSunjicTail(x, h, mu, sigmaDS, asym, tail)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigmaDS}{function full width at half maximum}

\item{asym}{function asymmetry value}

\item{tail}{amplitude of the spectral tail on the low BE (high KE) side}
}
\description{
DoniachSunjic function
Version of the Doniach-Sunjic function (see Wertheim PRB 25(3), 1987, (1982))
Corrected for an exponential decay (tail) on the low BE side
}
