% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{findYIndex}
\alias{findYIndex}
\title{findYIndex function utility}
\usage{
findYIndex(Y, Value, tolerance = 0.01)
}
\arguments{
\item{Y}{numeric vector}

\item{Value}{numeric value}

\item{tolerance}{numeric value, precision required to find out Y}
}
\value{
Returns Y index
}
\description{
finds the index of Y vector corresponding to a given Value
}
\examples{

\dontrun{
findYIndex(X, Value)
}

}
