% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitCompClass.r
\docType{class}
\name{fitComponents-class}
\alias{fitComponents-class}
\title{Class "fitComponents"}
\description{
A class that describes a fit component.  The class has the definition of all
the necessary informations needed to use a component algorithm with the
optimisation framework and the graphical user interface (but see Notes
below).
}
\note{
The goal is that the optimisation framework and the GUI code should get
all information about available baseline algorithms through a list of
\code{fitComponents} objects.  This will make it relatively simple to add new
baseline algorithms.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("fitComponents", ...)}.
}

\examples{

\dontrun{
showClass("fitComponents")
}
}
\author{
Roberto Canteri
}
\keyword{classes}
