% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Gauss}
\alias{Gauss}
\title{Gauss function}
\usage{
Gauss(x, h, mu, sigma)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}
}
\description{
Gauss function
}
