% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElements.r
\name{getElementValue}
\alias{getElementValue}
\title{Elements functions utility}
\usage{
getElementValue(element, orbital, analyzer = c("scienta", "kratos"),
  what = "RSF")
}
\arguments{
\item{element}{element symbol as character}

\item{orbital}{atom orbital}

\item{analyzer}{name of reference instrument dependent table of values}

\item{what}{name of value to get. Default \code{"RSF"}}
}
\value{
Return values named by \code{what}.
}
\description{
Unified functions to access and get values from the references files supplied
for Scienta and Kratos instruments. These function are internal functions and
not intended for the user.
}
\details{
Unified functions to acces and get values from the references files supplied
for Scienta and Kratos instruments.
}
