% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSGetHvalue.r
\name{GetHvalue}
\alias{GetHvalue}
\title{In interactive graphic panels the ordinate of the cursor positin is used to
compute the intensity of the selected fit function to add or to move.
see analysis or Move Component options.}
\usage{
GetHvalue(Object, Comp, FitFunct, A)
}
\arguments{
\item{Object}{the fitted CoreLine;}

\item{Comp}{the fitting component which is added or moved;}

\item{FitFunct}{thee fitting function corresponding to the component;}

\item{A}{amplitude of the fitting component obtained from the mouse locator}
}
\value{
Returns the component intensity.
}
\description{
In interactive graphic panels the ordinate of the cursor positin is used to
compute the intensity of the selected fit function to add or to move.
see analysis or Move Component options.
}
