% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{HillSigmoid}
\alias{HillSigmoid}
\title{HillSigmoid function}
\usage{
HillSigmoid(x, h, mu, pow, A, B)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{pow}{sigmoid decay rate}

\item{A}{Sigmoid upper limit}

\item{B}{Sigmoid lower limit}
}
\description{
HillSigmoid function
}
