% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{joinf}
\alias{joinf}
\title{Join factors}
\usage{
joinf(...)
}
\arguments{
\item{...}{One or more features to merge. 
Given as single vectors or data.frames. 
Accepted feature types are factor (preferred), booleans, integers (treated as categorical) or reals (which undergo automatic categorisation). 
\code{NA}s are not allowed.}
}
\value{
Joint factor, with levels \code{l1} to \code{l<n>}.
Vacant combinations are dropped.
}
\description{
Convenience function for joining factors.
}
\note{
You can pass a single vector to this function to see how praznik interprets it.
}
\examples{
joinf(c(T,F,T,F),c(T,T,F,F))
}
