% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Linear}
\alias{Linear}
\title{linear function}
\usage{
Linear(x, m, c, mu)
}
\arguments{
\item{x}{numeric vector}

\item{m}{slope}

\item{c}{constant value}

\item{mu}{function position}
}
\description{
linear function
}
