% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{MakeTable}
\alias{MakeTable}
\title{Font=GCourierNew is the CourierNew modified with symbols for drawing tab borders.
Provides commands to print borders and cell text alignment (left, center, right)
Borders are made using ascii characters}
\usage{
MakeTable(Type, txt, CellLength, align)
}
\arguments{
\item{Type}{string = BorderUp, BorderIn, BorderBottom, TableRow: prints border or table rows}

\item{txt}{array containing the txt of the table row}

\item{CellLength}{array containing the number of character of each table column}

\item{align}{string = left, center, right: text alignment of the table cells}
}
\value{
Returns the processed \code{object}.
}
\description{
Font=GCourierNew is the CourierNew modified with symbols for drawing tab borders.
Provides commands to print borders and cell text alignment (left, center, right)
Borders are made using ascii characters
}
\examples{

\dontrun{
  txt<-c("Components", "Area(cps)", "FWHM", "RSF", "BE(eV)", "TOT.(\%)")    #Voci Tabella
  CellLgth<-c(12, 15, 8, 7, 8, 9)
  cell<-MakeTable("MultiCell", txt,CellLength, "center")
}

}
