% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{MRMR}
\alias{MRMR}
\title{Minimum redundancy maximal relevancy filter}
\usage{
MRMR(X, Y, k = if (positive) ncol(X) else 3, positive = FALSE, threads = 0)
}
\arguments{
\item{X}{Attribute table, given as a data frame with either factors (preferred), booleans, integers (treated as categorical) or reals (which undergo automatic categorisation; see below for details).
Single vector will be interpreted as a data.frame with one column.
\code{NA}s are not allowed.}

\item{Y}{Decision attribute; should be given as a factor, but other options are accepted, exactly like for attributes.
\code{NA}s are not allowed.}

\item{k}{Number of attributes to select.
Must not exceed \code{ncol(X)}.}

\item{positive}{If true, algorithm won't return features with negative scores (i.e., with redundancy term higher than the relevance therm).
In that case, \code{k} controls the maximal number of returned features, and is set to `ncol(X)` by default.}

\item{threads}{Number of threads to use; default value, 0, means all available to OpenMP.}
}
\value{
A list with two elements: \code{selection}, a vector of indices of the selected features in the selection order, and \code{score}, a vector of corresponding feature scores.
Names of both vectors will correspond to the names of features in \code{X}.
Both vectors will be at most of a length \code{k}, as the selection may stop sooner, even during initial selection, in which case both vectors will be empty.
}
\description{
The method starts with a feature of a maximal mutual information with the decision \eqn{Y}.
Then, it greedily adds feature \eqn{X} with a maximal value of the following criterion:
\deqn{J(X)=I(X;Y)-\frac{1}{|S|}\sum_{W\in S} I(X;W),}
where \eqn{S} is the set of already selected features.
}
\note{
The method requires input to be discrete to use empirical estimators of distribution, and, consequently, information gain or entropy.
To allow smoother user experience, praznik automatically coerces non-factor vectors in inputs, which requires additional time, memory and may yield confusing results -- the best practice is to convert data to factors prior to feeding them in this function.
Real attributes are cut into about 10 equally-spaced bins, following the heuristic often used in literature.
Precise number of cuts depends on the number of objects; namely, it is \eqn{n/3}, but never less than 2 and never more than 10.
Integers (which technically are also numeric) are treated as categorical variables (for compatibility with similar software), so in a very different way -- one should be aware that an actually numeric attribute which happens to be an integer could be coerced into a \eqn{n}-level categorical, which would have a perfect mutual information score and would likely become a very disruptive false positive.
}
\examples{
data(MadelonD)
MRMR(MadelonD$X,MadelonD$Y,20)
}
\references{
"Feature Selection Based on Mutual Information: Criteria of Max-Dependency, Max-Relevance, and Min-Redundancy" H. Peng et al. IEEE Pattern Analysis and Machine Intelligence (PAMI) (2005)
}
