% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSnlsLM.r
\name{nlsModel}
\alias{nlsModel}
\title{function to perform the Lev. Marq. peak fitting}
\usage{
nlsModel(form, data, start, wts, Niter, upper = NULL)
}
\arguments{
\item{form}{expresson containing the fitting function to be evaluated}

\item{data}{the XPSCoreline to be fitted}

\item{start}{the starting estimates for the fitting function}

\item{wts}{optional numeric vector of weights. If present weighted least squares is performed}

\item{Niter}{max iteration allowed}

\item{upper}{upper,lower bounds for the estimates}
}
\description{
This is a modified version of the nlsLM() function
to avoid the frequent non-convergence problems arising
when multiple constraints are set on a given XPS CoreLine
}
