% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.vamas.r
\name{read.vamas}
\alias{read.vamas}
\title{read.vamas}
\usage{
read.vamas(file = NULL, debug = FALSE)
}
\arguments{
\item{file}{File}

\item{debug}{debug switch}
}
\description{
Read vamas format
}
\details{
Read vamas format
}
