% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.Genplot.r
\name{readGenplot}
\alias{readGenplot}
\title{read.vamas}
\usage{
readGenplot(Object, file = NULL, debug = FALSE)
}
\arguments{
\item{Object}{an XPSSample object where to put the read data}

\item{file}{File}

\item{debug}{debug switch}
}
\description{
Read Genplot RPL data
}
\details{
Read Genplot analyzed spectra in RPL format
a RPL folder containing analyzed data must be present
}
