% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{Shirley2P}
\alias{Shirley2P}
\title{Shirley2P  is a two parameter Bs, Cs expressely evaluated for the Shirley background
Shirley Cross Section described by: B*T/(C + T^2)     T=energy loss
see J. Vegh, Surface Science 563 (2004) 183,    J. Vegh  J.Electr.Spect.Rel.Phen. 151 (2006) 159}
\usage{
Shirley2P(object, limits)
}
\arguments{
\item{object}{XPSSample object}

\item{limits}{limits of the XPSSample object}
}
\value{
returns the Two_parameter Shirley background.
}
\description{
Shirley2P  is a two parameter Bs, Cs expressely evaluated for the Shirley background
Shirley Cross Section described by: B*T/(C + T^2)     T=energy loss
see J. Vegh, Surface Science 563 (2004) 183,    J. Vegh  J.Electr.Spect.Rel.Phen. 151 (2006) 159
}
