% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{Shirley3P}
\alias{Shirley3P}
\title{Shirley3p is a three parameters Bs=30eV^2, Cs=750eV^2, Ds=0.75 are expressely evaluated for the Shirley background
Shirley Cross Section described by: B*T/(C + T^2)*(1-exp(-Ds*T))     T=energy loss
see J. Vegh  J.Electr.Spect.Rel.Phen. 151 (2006) 159}
\usage{
Shirley3P(object, Wgt, limits)
}
\arguments{
\item{object}{XPSSample object}

\item{Wgt}{XPSSample object}

\item{limits}{limits of the XPSSample object}
}
\value{
returns the 3-parameter Shirley background.
}
\description{
Shirley3p is a three parameters Bs=30eV^2, Cs=750eV^2, Ds=0.75 are expressely evaluated for the Shirley background
Shirley Cross Section described by: B*T/(C + T^2)*(1-exp(-Ds*T))     T=energy loss
see J. Vegh  J.Electr.Spect.Rel.Phen. 151 (2006) 159
}
