% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElements.r
\name{showTableElement}
\alias{showTableElement}
\title{Elements functions utility}
\usage{
showTableElement(element, analyzer = c("scienta", "kratos"))
}
\arguments{
\item{element}{element symbol as character}

\item{analyzer}{name of reference list}
}
\value{
Return values named by \code{what}.
}
\description{
Unified functions to acces and get values from the references files supplied
for Scienta and Kratos instruments. These function are internal functions and
not intended for the user.
}
\details{
Unified functions to acces and get values from the references files supplied
for Scienta and Kratos instruments.
}
