% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid function}
\usage{
Sigmoid(x, h, mu, k, c)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{k}{sigmoid decay rate}

\item{c}{constant value}
}
\description{
Sigmoid function
}
