% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{[,XPSSample-method}
\alias{[,XPSSample-method}
\title{Get subset of XPSSample}
\usage{
\S4method{[}{XPSSample}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{the \code{XPS} object of class XPSSample to be generated}

\item{i}{numeric parameter corresponding to the number of spectra}

\item{j}{dummy numeric parameter}

\item{drop}{logical by default FALSE}
}
\value{
XPSSample
}
\description{
Get subset of XPSSample
}
