% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomBaseline.r
\name{Tougaard4P}
\alias{Tougaard4P}
\title{Tougaard4P is the four parameters B, C, C' and D Tougaard background
Universal Cross Section described by: B*T/[(C + C'*T^2)^2 + D*T^2]    T=energy loss
see R. Hesse et al., Surf. Interface Anal. (2011), 43, 1514}
\usage{
Tougaard4P(object, C1, limits)
}
\arguments{
\item{object}{XPSSample object}

\item{C1}{numeric is a parameter value}

\item{limits}{limits of the XPSSample object}
}
\value{
Return the 4parameter Tougaard background.
}
\description{
Tougaard4P is the four parameters B, C, C' and D Tougaard background
Universal Cross Section described by: B*T/[(C + C'*T^2)^2 + D*T^2]    T=energy loss
see R. Hesse et al., Surf. Interface Anal. (2011), 43, 1514
}
