% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{Voigt}
\alias{Voigt}
\title{Voigt function}
\usage{
Voigt(x, h, mu, sigma, lg)
}
\arguments{
\item{x}{numeric vector}

\item{h}{function amplitude}

\item{mu}{position of function center}

\item{sigma}{function full width at half maximum}

\item{lg}{mix Gauss-Lorentz}
}
\description{
Voigt function
}
