% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volc.loc.r
\docType{data}
\name{volc.loc}
\alias{volc.loc}
\title{Dataset: location on the volcano Maunga Whau (Auckland, New Zealand)}
\format{SpatialPointsDataFrame}
\usage{
data(volc.loc)
}
\description{
A SpatialPointsDataFrame representing a spot on the volcano Maunga Whau (Auckland, New Zealand).
}
\keyword{datasets}
