% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSAnnotateGUI.r
\name{XPSAnnotate}
\alias{XPSAnnotate}
\title{Annotate: graphic function to add text to plots}
\usage{
XPSAnnotate()
}
\description{
Add text to spectra plotted using (\code{matplot})base function.
No parameters are passed to the this function
}
\examples{

\dontrun{
annotate()
}

}
\seealso{
\link{matplot}, \link{plot}
}
