% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSapplyshift,XPSSample-method}
\alias{XPSapplyshift,XPSSample-method}
\title{Apply Energy Shift to the XPSSample}
\usage{
\S4method{XPSapplyshift}{XPSSample}(object, shift = NULL)
}
\arguments{
\item{object}{XPSSample}

\item{shift}{shift value}
}
\description{
Apply Energy Shift Method for XPSSample
}
\details{
Apply the shift value to the Binding axis. If \code{shift} is NULL then 
the x-axis will be reset to the original values.
}
