% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSapplyshift}
\alias{XPSapplyshift}
\title{Apply Shift to the XPSCoreLine}
\usage{
XPSapplyshift(object, shift = NULL)
}
\arguments{
\item{object}{XPSCoreLine}

\item{shift}{X-shift value}
}
\description{
Apply Shift Method for XPSCoreLine
}
\details{
Apply the shift value to the X axis. If \code{shift} is NULL then
the x-axis will be reset to the original values.
}
