% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSbaseline}
\alias{XPSbaseline}
\title{Baseline definition for XPSCoreLine object.}
\usage{
XPSbaseline(object, bgtype = c("shirley", "polynomial", "spline"),
  deg = NULL, Wgt = NULL, splinePoints = list(x = NULL, y = NULL),
  ...)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{bgtype}{the baseline type. For linear type \code{"linear"}. See
\link[baseline:baseline]{baseline} for other baseline types.}

\item{deg}{degree of the polynomial background}

\item{Wgt}{LinearPolynomial weigth in LPShirley}

\item{splinePoints}{numeric vector containing the points which will be connected by the spline}

\item{...}{other parameters to the \link[baseline:baseline]{baseline}
functions. Use only for baseline function different from \code{"shirley"} or
\code{"linear"}.}
}
\value{
The Object slot \code{Baseline} will be filled with the result and it
will be displyed. The baseline function return an object of class
\code{baseline}. The x is the same as \code{RegionToFit} x coord. The y coord
are the baseline values. The user who wish a particular baseline not yet
implemented, could do the calculation without the use of this
\code{XPSbaseline} and fill manually the \code{Baseline} slot.
}
\description{
Calculate the baseline for XPSCoreLine. This function acts as wrapper for the
\link[baseline:baseline]{baseline} function. There are two tipical shape for
baseline: \code{"shirley"} and \code{"linear"} but the package
\link[baseline:baseline]{baseline} has many other shapes implemented.
}
\details{
The function \code{baseline.shirley} implements the \code{shirley baseline}
and adds it to the package \code{baseline}.  It is an iterative algorithm.
The iteration stops when the deviation between two consequent iteration is
lower than \code{err} or when the max number of iterations \code{maxit} is
reached. Normally it should not be used directly by the user.
}
\examples{

\dontrun{
test[["C1s"]] <- XPSbaseline(test[["C1s"]], "linear")
}

}
\seealso{
\link[baseline:baseline]{baseline}
}
