% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPScalc}
\alias{XPScalc}
\title{Function to compute the integral intensity for a given XPS-CoreLine.}
\usage{
XPScalc(object, table = TRUE)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{table}{Print table Logic}
}
\value{
An ASCII file for each XPSCoreLine.
}
\description{
Calculation of Integral Intensity of fitting components of a XPSCoreLine.
}
\seealso{
\link{write.table}, \link{write.csv}, \link{write.csv2}
}
