% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSConstraintsGUI.r
\name{XPSConstraints}
\alias{XPSConstraints}
\title{GUI to add a fit constraints to the fitting functions of a XPSCoreLine}
\usage{
XPSConstraints()
}
\description{
Add constraints among fit components defined for a XPSCoreLine object. 
This function is called after the definition of the baseline (\code{XPSbaseline}) 
and fitting components (\code{XPSaddFitComponent}). No parameters are passed to
this function
}
\examples{

\dontrun{
XPSConstraints()
}

}
\seealso{
\link{XPSconstrain}, \link{XPSbaseline}, \link{XPSaddFitComponent}, \link{XPSfitAlgorithms}
}
