% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{class}
\name{XPSCoreLine-class}
\alias{XPSCoreLine-class}
\title{Class "XPSCoreLine"}
\description{
The package provides a class for XPS spectra (class \code{XPSCoreLine}) and
lists of such objects (class \code{XPSSample}). \code{XPSCoreLine} are values
pairs stored in a \code{list} and several additional parameters stored in
slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("XPSCoreLine", ...)}.
}

\examples{

\dontrun{
test <- new("XPSCoreLine", Info="test", units=c("Binding [eV]", "Counts"))
}

}
\keyword{classes}
