% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCoreLineList.r
\name{XPSCoreLineList}
\alias{XPSCoreLineList}
\title{Make the list of CoreLines of the loaded XPS-Sample}
\usage{
XPSCoreLineList()
}
\description{
The list of corelines which are contained in the loaded XPS-Sample
are displayed by this function. No parameters are passed to this function.
}
\examples{

\dontrun{
XPSCoreLineList()
}

}
