% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCustomPlotGUI.r
\name{XPSCustomPlot}
\alias{XPSCustomPlot}
\title{Function to generate personalized plot}
\usage{
XPSCustomPlot()
}
\description{
This GUI allows setting the various parameters genereally used for
replotting data. Through a user friendly interface it is possible to
set colors, lines or symbols, their weight, and annotate the plot.
This Gui is based on the Lattice package.
No parameters are passed to this function.
}
\examples{

\dontrun{
XPSCustomPlot()
}

}
\seealso{
\link{Lattice}
}
