% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSDiffGUI.r
\name{XPSDiff}
\alias{XPSDiff}
\title{Differentiate is a function to compute the derivative of the XPSSpectra.}
\usage{
XPSDiff()
}
\description{
Differentiate is a function to compute the derivative of the XPSSpectra.
}
\examples{

\dontrun{
XPSDiff()
}

}
