% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSElemTabGUI.r
\name{XPSElemTab}
\alias{XPSElemTab}
\title{CoreLine and Auger Transition tables}
\usage{
XPSElemTab()
}
\description{
GUI to help identification of elements in Survey spectra.
}
\examples{

\dontrun{
XPSElemTab()
}

}
