% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSEshiftGUI.r
\name{XPSEshift}
\alias{XPSEshift}
\title{XPSEShift function utility}
\usage{
XPSEshift()
}
\description{
Unified functions to alligne the energy scale through a reference peak
As reference peaks C1s from hydrocarbons at BE=285eV or Au4f 7/2 at BE=84eV
are generally chosen.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSEshift()
}

}
