% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSExportAsciiGUI.r
\name{XPSExportAscii}
\alias{XPSExportAscii}
\title{GUI for the selection of the more convenient style to export all the XPSCoreLine
in a XPSSample in anASCII text file.}
\usage{
XPSExportAscii()
}
\description{
Export all the XPSCoreLine in a XPSSample in ASCII format.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSexportASCII()
}

}
