% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSExtractGUI.r
\name{XPSExtractGUI}
\alias{XPSExtractGUI}
\title{Extract a portion of spectrum from a XPS survey}
\usage{
XPSExtractGUI(Object, coreline, plot_win = winsize)
}
\arguments{
\item{Object}{XPSSample object name}

\item{coreline}{index identifying the current XPS CoreLine}

\item{plot_win}{size of the graphical Window used}
}
\value{
returns the \code{object} with a coreline added.
}
\description{
Extract a portion (abscissa) from a XPS survey in a XPSSample. It use a GUI
to interact with the x-axis and to get the \code{Symbol} of the new
XPSCoreLine
}
\examples{

\dontrun{
XPSextractGUI(SampData, index, 1.6)
}

}
