% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFilterGUI.r
\name{XPSFilter}
\alias{XPSFilter}
\title{Smoothing functions to remove noise from spectra.}
\usage{
XPSFilter()
}
\description{
XPSFilter contains a list of different filters to remove noise from spectra.
}
\details{
In Sawitzy Golay filter at point i is a weighted average of data i-n... i ... i+n
In Autoregressive filters the output variable depends linearly on its own previous values. 
In the Moving Average filters the output at point i is the average of data i-n ... i ... i+n .
The FFT filter applys the FFT transform to perform filtering.
The Wavelets fileter uses the wavelets to perform filtering.
The FIRfilter is a Finite Impulse Response with a zero distortion.
The Butterworth is an Infinite Impulse Response filter.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSFilter()
}

}
