% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitInfoGUI.r
\name{XPSFitInfo}
\alias{XPSFitInfo}
\title{Provides information about a Core Line fit}
\usage{
XPSFitInfo()
}
\description{
Makes the computation of the integral intensity of each of the fitting components
of a given coreline. The sum of the component integral intensities =100% (it is the
best fit integral intensity). No parameters are passed to this function.
}
\examples{

\dontrun{
XPSFitInfo()
}

}
