% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitLM.r
\name{XPSFitLM}
\alias{XPSFitLM}
\title{Perform the fit of a Core Line}
\usage{
XPSFitLM(Object, plt = TRUE, ...)
}
\arguments{
\item{Object}{XPSCoreLine object}

\item{plt}{logical if TRUE residual plot is drawn}

\item{\dots}{further parameters to the fitting function}
}
\value{
The Object slot \code{FitLM} will be filled with the result of the
calculation. All the values of the components will be update and the result
will be displyed.
}
\description{
Calculate the fit of a XPSCoreLine object. The \code{XPSFitLM} performs the
fit and stop. This function is the main function for performing the fit. The
\code{XPSfit} is the main function which calls the \code{XPSFitLM} and
displays the result.
}
\examples{

\dontrun{
showListParam(c(1271, 285, 1.4), 3, c("h", "mu", "sigma" ) )
}

}
\seealso{
\link{nlsLM}
}
