% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitParamGUI.r
\name{XPSFitParam}
\alias{XPSFitParam}
\title{XPSFitParam Gui to set fit param MaXiTERATION, TOLERANCE, MINFACTOR, when NLSfit error occurs}
\usage{
XPSFitParam(ParamNames, ParamValues)
}
\arguments{
\item{ParamNames}{c("maxiteration", "tolerance", "minfactor") the set of parameter to work with}

\item{ParamValues}{the correspondent values}
}
\description{
XPSFitParam Gui to set fit param MaXiTERATION, TOLERANCE, MINFACTOR, when NLSfit error occurs
}
\examples{
\dontrun{
XPSFitParam(c("maxiteration", "tolerance", "minfactor"), c(1000, 1e-5, 1e-3))
}

}
