% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitParamInfoGUI.r
\name{XPSFitParamInfo}
\alias{XPSFitParamInfo}
\title{Provides the list of components and the relative fitting parameters}
\usage{
XPSFitParamInfo()
}
\description{
Provides the list of components and the relative fitting parameters
for a selected XPSCoreLine. No parameters are passed to this function
}
\examples{

\dontrun{
XPSFitParamInfo()
}

}
