% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFNameList.r
\name{XPSFNameList}
\alias{XPSFNameList}
\title{Provide a list of the XPS-Sample names loaded}
\usage{
XPSFNameList(warn = TRUE)
}
\arguments{
\item{warn}{logical enables warning messages}
}
\description{
Provide a list of names (strings) relative to the XPS-Samples
loaded in the XPS-Analysis software. No parameters are passed to this function.
}
\examples{

\dontrun{
XPSFNameList()
}

}
