% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSLatticeAnnotateGUI.r
\name{XPSLattAnnotate}
\alias{XPSLattAnnotate}
\title{Adds text to a plot produced using the Lattice package
This function is called by (\code{XPSOverlay}) and (\code{XPSCustomPlot})}
\usage{
XPSLattAnnotate(Xlim, Ylim)
}
\arguments{
\item{Xlim}{the limits of the X axis}

\item{Ylim}{the limits of the Y axis}
}
\description{
Adds text to a plot produced using the Lattice package
This function is called by (\code{XPSOverlay}) and (\code{XPSCustomPlot})
}
\examples{

\dontrun{
XPSLattAnnotate(Xlim, Ylim)
}

}
