% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSModFitGUI.r
\name{XPSModFit}
\alias{XPSModFit}
\title{Perform the fit of a Core Line}
\usage{
XPSModFit(Object, ...)
}
\arguments{
\item{Object}{XPSCoreLine object}

\item{\dots}{further parameters to the fitting function}
}
\value{
The Object slot \code{XPSModFit} will be filled with the result of the
calculation. All the values of the components will be update and the result
will be displyed.
}
\description{
Provides a userfriendly interface to select a fitting algorithms to
fit an XPSCoreline. Fitting algorithms are:
  Classic algorithms: Levenberg-Marquardt, Newton and Port based on the minimization 
  of the Squares of the Differences and handled in a more robust way (although slower)
  compared to (\code{nlsLM}) function.
  Conjugate-Gradient Algorithms: General-purpose optimization based on Nelder-Mead, 
  quasi-Newton and conjugate-gradient algorithms.
  Pseudo Algorithms: random based algorithm (rather slow) but ensure the convergence.
}
\examples{

\dontrun{
SampData[["C1s"]] <- XPSModFit(SampData[["C1s"]])
}

}
\seealso{
\link{nlsLM}
}
