% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSMoveBaseLineGUI.r
\name{XPSMoveBaseLine}
\alias{XPSMoveBaseLine}
\title{Modifies the BaseLine level and limits of a given Coreline}
\usage{
XPSMoveBaseLine()
}
\description{
Function to modify the BaseLine ends and level for a given CoreLine.
No parameters are passed to this function.
}
\examples{

\dontrun{
XPSMoveBaseLine()
}

}
