% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSMoveCompGUI.r
\name{XPSMoveComp}
\alias{XPSMoveComp}
\title{Function to modify component position and intensity in a fit}
\usage{
XPSMoveComp(XPSSample, Indx, hscale)
}
\arguments{
\item{XPSSample}{XPSSample object}

\item{Indx}{index identifyinf the CoreLine to work on}

\item{hscale}{dimension of the plotting window}
}
\description{
Provides a userfriendly interface change position and intensity of each
individual fitting component of a selected XPSCoreline. Changes are saved in the 
.GlobalEnv main software memory
}
\examples{

\dontrun{
XPSMoveComp(SampData, 2, "medium")
}

}
