% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSOverlayEngine.r
\name{XPSovEngine}
\alias{XPSovEngine}
\title{XPSOverlay Engine is the software powering the XPSOverlayGUI.}
\usage{
XPSovEngine(PlotParameters, Plot_Args, AutoKey_Args, SelectedNames, Xlim,
  Ylim)
}
\arguments{
\item{PlotParameters}{the plot parameters asociated to the XPSOverlayGUI options;}

\item{Plot_Args}{list of plot options;}

\item{AutoKey_Args}{list of options for annotation;}

\item{SelectedNames}{list containing the XPSSample names and the Corelines to be plotted;}

\item{Xlim}{Xrange of the data to be plotted;}

\item{Ylim}{Yrange of the data to be plotted;}
}
\value{
Return c(Xlim, Ylim) eventually modified
}
\description{
XPSOverlay Engine is the software powering the XPSOverlayGUI.
}
