% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSPreferencesGUI.r
\name{XPSPreferences}
\alias{XPSPreferences}
\title{To select a dimensions of the graphic window depending on the dimensions of the screen used}
\usage{
XPSPreferences()
}
\description{
To select a dimensions of the graphic window depending on the dimensions of the screen used
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSetWinSize()
}

}
