% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSProcess.r
\name{XPSprocess}
\alias{XPSprocess}
\title{GUI to analyze XPS-Corelines}
\usage{
XPSprocess(Object, Object_name = as.character(substitute(Object)),
  TKwin = winsize)
}
\arguments{
\item{Object}{XPSSample object}

\item{Object_name}{XPSSample name as character}

\item{TKwin}{Graphical windows dimensions}
}
\value{
Return the processed \code{object}.
}
\description{
Interactive GUI to add BaseLines and Fitting components to a CoreLine
}
