% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSProcessCoreLineGUI.r
\name{XPSProcessCoreLine}
\alias{XPSProcessCoreLine}
\title{Add a CoreLine (with fit if existing)to an XPS-Sample, or just BaseLine and Fit
or remove a Coreline from a XPS-Sample}
\usage{
XPSProcessCoreLine()
}
\description{
Add a Coreline from a source XPS-Sample into a destination XPS-Sample or
Add a just a BaseLine and Fit from a source XPS-Sample into a destination XPS-Sample or
remove a CoreLine from a XPS-Sample. No parameters are passed to this function
}
\examples{

\dontrun{
ProcessCoreLine()
}

}
