% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSQuantifyGUI.r
\name{XPSQuant}
\alias{XPSQuant}
\title{Perform the elemental quantification for a selected XPS-Sample}
\usage{
XPSQuant()
}
\description{
Provides a userfriendly interface with the list of Corelines of the selected XPS-Sample
Each Coreline can be count/omit from the elemental quantification.
If peak fitting is present also each of the fitting component can be count/omit from the
elemental quantification. 
Finally also relative RSF of the coreline or individual fitting components can be modified.
No parametrs are passed to this function.
}
\examples{

\dontrun{
XPSQuant()
}

}
