% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.Oldscienta.r
\name{XPSRead.Oldscienta}
\alias{XPSRead.Oldscienta}
\title{Read XPS data from old-DOS-Scienta format}
\usage{
XPSRead.Oldscienta(filename = NULL, project = "", experiment = "",
  out = c("list"))
}
\arguments{
\item{filename}{file name or list of file names}

\item{project}{comment}

\item{experiment}{comment}

\item{out}{output class type default list()}
}
\value{
list object class
}
\description{
Reads XPS data 'old-Scienta' format Data.
}
\examples{

\dontrun{
pk1 <- read.xps("/Volumes/GoogleDrive/Il mio Drive/R/OldScientaFiles/Specific/DLC.48")
}

}
