% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRead.r
\name{XPSread}
\alias{XPSread}
\title{Read '.pxt' and '.vms' format Data}
\usage{
XPSread(file = NULL, Genplot = TRUE, ...)
}
\arguments{
\item{file}{file name}

\item{Genplot}{flag to read genplot .RPL files for 'Scienta' format files}

\item{\dots}{further parameters to \code{read.scienta} and \code{read.vamas}}
}
\value{
returns an object of class XPSSample
}
\description{
Reads a file and returns an XPSSpectra class item. In the case the file
extension is \code{.pxt} and the option \code{Genplot} is \code{TRUE}, then a
Genplot \code{RPL} file with the same name will be searched in a folder
\code{RPL} and it will be sourced if founded.
}
\note{
\code{read.scienta} \code{read.vamas} and \code{readGenplot} 
must not used by the end-user
}
\examples{

\dontrun{
SampData1 <- XPSread("SampData1.pxt")
}

\dontrun{
SampData2 <- XPSread("SampData2.vms")
}

}
