% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSReportGUI.r
\name{XPSReport}
\alias{XPSReport}
\title{Provides information about a Core Line fit}
\usage{
XPSReport()
}
\description{
Makes the computation of the integral intensity of each of the fitting components
of a given coreline. The sum of the component integral intensities =100% (it is the
best fit integral intensity). No parameters are passed to this function.
}
\examples{

\dontrun{
XPSReport()
}

}
