% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSResetAnalysisGUI.r
\name{XPSResetAnalysis}
\alias{XPSResetAnalysis}
\title{Resets a XPS-Coreline processing}
\usage{
XPSResetAnalysis()
}
\description{
Function to delete individual elements of a Coreline analysis:
resets only the Best Fit;
resets an single or all the Fitting Components;
resets the BaseLine (i.e. eliminates alsio the Fit);
resets ALL alle the Coreline analysis.
no parameters are passed to this function
}
\examples{

\dontrun{
XPSResetAnalysis()
}

}
