% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{class}
\name{XPSSample-class}
\alias{XPSSample-class}
\title{Class "XPSSample"}
\description{
The package provides classes for XPS spectra (class \code{XPSSample}) and
collections of such lists (class \code{XPSCoreLine}). \code{XPSSample} are
mass-intensity value pairs stored in a two column \code{list} and several
additional parameters stored in slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("XPSSample", data, Project, Sample, Comments, User, Filename)}.
}

\examples{

\dontrun{
SiOx <- new("XPSSample",info="test",experiment="Si-Ti Waveguides")
}
}
\keyword{classes}
