% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSampleInfoGUI.r
\name{XPSSampleInfo}
\alias{XPSSampleInfo}
\title{XPSSampleInfo to show/modify XPSSample INFOs is saved during acquisition}
\usage{
XPSSampleInfo()
}
\description{
XPSSampleInfo to show/modify XPSSample INFOs is saved during acquisition
}
\examples{

\dontrun{
XPSSampleInfo()
}

}
