% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSaveDataGUI.r
\name{XPSSaveData}
\alias{XPSSaveData}
\title{To Save data in the Hard Disk}
\usage{
XPSSaveData()
}
\description{
Provides a userfriendly interface to select a FileName and Directory
where to save the analyzed XPS-Sample data.
Analyzed data by default have extension .Rdata.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSaveData()
}

}
