% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetCoreLineGUI.r
\name{XPSSetCoreLine}
\alias{XPSSetCoreLine}
\title{To select a Core Line}
\usage{
XPSSetCoreLine()
}
\description{
The list of corelines of a give XPS-Sample are presented for selection
no parameters are passed to this function
}
\examples{

\dontrun{
XPSSetCoreLine()
}

}
