% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetFNameCLineGUI.r
\name{XPSSetFNameCLine}
\alias{XPSSetFNameCLine}
\title{To select a XPS-Sample and a Core Line}
\usage{
XPSSetFNameCLine()
}
\description{
The list of XPS-Samples loaded is presented for selection.
After selection of the XPS-Sample the list of correspondent Corelines is available for selection.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSetFNameCLine()
}

}
