% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetGraphDevGUI.r
\name{XPSSetGraphDev}
\alias{XPSSetGraphDev}
\title{To select the kind of graphic device compatible with the operating system in use}
\usage{
XPSSetGraphDev()
}
\description{
To select the kind of graphic device compatible with the operating system in use
Also a list of graphic formats is provided to save the content of the current
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSetGraphDev()
}

}
