% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSsetRegionToFit}
\alias{XPSsetRegionToFit}
\title{Set Region to Fit}
\usage{
XPSsetRegionToFit(object, limits, ...)
}
\arguments{
\item{object}{XPSCoreLine object}

\item{limits}{list with x,y values to limit the region.}

\item{...}{further parameters to the plot function}
}
\value{
The limits usually will be set with the cursor on the plot of the
curve. This step is normally do when a \code{Baseline} is required. Then the
step of definition of the \code{RegionToFit} is accoplished. The portion of
the original curve as a list of (x,y) will be filled into the slot
\code{RegionToFit}.  %Please note that the \code{RegionToFit} definition
also include the transformation axis of the original curve, to "Binding
Energy" for the x-axis and "counts per second (cps)" for the y-axis.
}
\description{
Definition of the portion (usually on the x-axis) of the original curve for
further processing
}
\examples{

\dontrun{
test[["C1s"]] <- XPSsetRegionToFit(test[["C1s"]])
}

}
\seealso{
\link{XPSbaseline}
}
