% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetWDGUI.r
\name{XPSSetWD}
\alias{XPSSetWD}
\title{To set the default working dir if it is not defined}
\usage{
XPSSetWD()
}
\description{
The new default working dir will be saved in the XPSSettings.ini
which uincludes also the user preferences.
}
\examples{

\dontrun{
XPSSetWD()
}

}
