% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSetWinSizeGUI.r
\name{XPSSetWinSize}
\alias{XPSSetWinSize}
\title{To select a dimensions of the graphic window depending on the dimensions of the screen used}
\usage{
XPSSetWinSize()
}
\description{
To select a dimensions of the graphic window depending on the dimensions of the screen used
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSetWinSize()
}

}
