% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSpectList.r
\name{XPSSpectList}
\alias{XPSSpectList}
\title{Show the list of spctra related to an XPSSample identified by its Spectral ID}
\usage{
XPSSpectList(activeFName, noIdx = FALSE)
}
\arguments{
\item{activeFName}{the name (string) of the selected XPSSample}

\item{noIdx}{logical if TRUE a numerical index is added to any CoreLine name}
}
\description{
Show the list of spctra related to an XPSSample identified by its Spectral ID
}
\examples{

\dontrun{
XPSSampleInfo(activeFName, noIdx=FALSE)
}

}
