% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSpectNameChangeGUI.r
\name{XPSSpectNameChange}
\alias{XPSSpectNameChange}
\title{GUI to change the name of the XPSSample and/or the name associated to corelines}
\usage{
XPSSpectNameChange()
}
\description{
GUI to change the name of the XPSSample and/or the name associated to corelines
}
\examples{

\dontrun{
XPSSpectNameChange()
}

}
