% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSplitGUI.r
\name{XPSSplit}
\alias{XPSSplit}
\title{To select split multiple acquisitions performed with the ScientaGammadata instrument}
\usage{
XPSSplit()
}
\description{
Acquisitions on multiple samples may be included in a single .PXT file
This function allows splitting spectra corresponding to each sample in 
individual files.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSplit()
}

}
