% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSprucingGUI.r
\name{XPSSprucingGUI}
\alias{XPSSprucingGUI}
\title{Sprucing XPSSample datas}
\usage{
XPSSprucingGUI(Object, coreline, plot_win = 1.3)
}
\arguments{
\item{Object}{XPSSample object name}

\item{coreline}{numeric index identifying the current XPS CoreLine}

\item{plot_win}{Window size default=1.3}
}
\value{
Returns the \code{Object} with a coreline added.
}
\description{
GUI to correct original XPS spectral data
}
\examples{

\dontrun{
XPSextractGUI(SampData, Index, "medium")
}

}
