% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSurElmtIdentifyGUI.r
\name{XPSSurveyElementIdentify}
\alias{XPSSurveyElementIdentify}
\title{Identify Elements on the Survey}
\usage{
XPSSurveyElementIdentify(CurrentXPSSample, SpectIdx)
}
\arguments{
\item{CurrentXPSSample}{XPSSample name}

\item{SpectIdx}{XPSCoreline index}
}
\description{
GUI to Identify Elements on the Survey.
}
