% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSVmsCorrGUI.r
\name{XPSVmsCorr}
\alias{XPSVmsCorr}
\title{Corrects XPSSample Vamas_type data for the analyzer transmission}
\usage{
XPSVmsCorr()
}
\description{
This routine divides raw data intensity for the value of the analyzer transmission
The routine applies the correction to the selected coreline or to all the XPSSample
spectra, baselines and fit components when present.
}
\examples{

\dontrun{
XPSVmsCorr()
}

}
