% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSZoomCurGUI.r
\name{XPSZoomCur}
\alias{XPSZoomCur}
\title{To zoom a plotted spectrum}
\usage{
XPSZoomCur()
}
\description{
To zoom a plotted spectrum.
No parameters are passed to this function
}
\examples{

\dontrun{
XPSZoomCur()
}

}
