% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSurElmtIdentifyGUI.r
\name{XPSSurveyElementIdentify}
\alias{XPSSurveyElementIdentify}
\title{Identify Peaks in a Survey and suggests elements which can be assigned 
to the identifie spectral features}
\usage{
XPSSurveyElementIdentify()
}
\description{
GUI to Identify Elements on the Survey.
}
\examples{

\dontrun{
XPSSurveyElementIdentify()
}

}
